exercise 1 - added checks

+++ VehicleManagement/Interfaces/IUserInputService.cs
@@ -4,7 +4,7 @@
     {
         public void PrintMenu();
         public void InputVehicleComponents(out string brand, out string model, out int year);
-        public int InputTruckComponent();
+        public decimal InputTruckComponent();
         public bool InputMotorcycleComponent();
         public int InputElectricRange();
         public int InputCarComponent();
